requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flex==6.14.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonpointer==2.0
jwt==1.2.0
MarkupSafe==1.1.1
packaging==20.9
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3987==1.3.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.26.3
validate-email==1.3
whitenoise==5.2.0
